Rename : 백터 아이콘 패키지 이동
1. ui / home / component -> ui / icon
