+++ app/src/main/java/com/goody/dalda/ui/home/component/AlcoholCategory.kt
@@ -11,7 +11,6 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.width
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -21,7 +20,8 @@ import androidx.compose.ui.unit.dp
 import com.goody.dalda.data.AlcoholCategoryStatus
 import com.goody.dalda.data.AlcoholType
 import com.goody.dalda.ui.component.AutoResizedText
-import com.goody.dalda.ui.home.component.iconpack.IcComingSoon
+import com.goody.dalda.ui.icon.iconpack.IcComingSoon
+import com.goody.dalda.ui.icon.IconPack
 import com.goody.dalda.ui.theme.DaldaTextStyle
 
 @OptIn(ExperimentalLayoutApi::class)
