+++ app/src/main/java/com/goody/dalda/ui/search/SearchScreen.kt
@@ -19,8 +19,8 @@ import com.goody.dalda.data.AlcoholData
 import com.goody.dalda.data.AlcoholType
 import com.goody.dalda.ui.AppPaddingSize
 import com.goody.dalda.ui.component.SearchBarComponent
-import com.goody.dalda.ui.home.component.IconPack
-import com.goody.dalda.ui.home.component.iconpack.IcCamera
+import com.goody.dalda.ui.icon.IconPack
+import com.goody.dalda.ui.icon.iconpack.IcCamera
 import com.goody.dalda.ui.search.component.RecommendAlcoholList
 import com.goody.dalda.ui.search.component.ResentSearch
